Refactor backend tests into modular files and update README
