+++ .github/workflows/main.yml
@@ -52,7 +52,7 @@ jobs:
         run: |
           source env/bin/activate
           cd ${{github.workspace}}/backend
-          pytest -s websocket_unit_test.py
+          pytest -s tests
 
   
   lint:
